requirements.txt
pyyaml
psutil
requests

numpy
opencv-python
face-recognition

paho-mqtt
